+++ src/integrationTest/java/uk/gov/hmcts/darts/event/service/impl/StandardEventHandlerTest.java
@@ -320,7 +320,7 @@ class StandardEventHandlerTest extends HandlerTestData {
             .containsOnly(idFrom(persistedEvents));
     }
 
-    private Integer idFrom(List<EventEntity> eventEntities) {
+    private Long idFrom(List<EventEntity> eventEntities) {
         return eventEntities.getFirst().getId();
     }
 }
